Modificación metodos para buscar por id, actualizar y eliminar

+++ src/main/java/com/example/demo/service/serviceImpl/ArticulosServiceImpl.java
@@ -2,7 +2,12 @@ package com.example.demo.service.serviceImpl;
 
 import com.example.demo.dao.ArticulosDAO;
 import com.example.demo.model.Articulos;
+import com.example.demo.model.Categoria;
+import com.example.demo.model.Estado;
+import com.example.demo.model.dto.ArticulosDTO;
 import com.example.demo.service.ArticulosService;
+import com.example.demo.service.CategoriaService;
+import com.example.demo.service.EstadoService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Service;
@@ -15,8 +20,14 @@ public class ArticulosServiceImpl implements ArticulosService {
     @Autowired
     ArticulosDAO articulosDAO;
 
+    @Autowired
+    EstadoService estadoService;
+
+    @Autowired
+    CategoriaService categoriaService;
+
     @Override
-    public String save(Articulos articulos) throws Exception {
+    public String registrar(Articulos articulos) throws Exception {
         validaciones(articulos);
         articulosDAO.save(articulos);
         return "Se creo correctamente el articulo";
@@ -24,19 +35,46 @@ public class ArticulosServiceImpl implements ArticulosService {
     }
 
     @Override
-    public String update(Articulos articulos) throws Exception {
-        validaciones(articulos);
-        articulosDAO.save(articulos);
+    public String actualizar(ArticulosDTO articulosDTO) throws Exception {
+        Articulos articulo = null;
+        Estado idEstado = null;
+        Categoria idCategoria = null;
+        //Todo: Faltan validaciones
+        articulo = articulosDAO.findById(articulosDTO.getIdArticulo()).orElse(null);
+        articulo.setPrecio(articulosDTO.getPrecio());
+        articulo.setImagen(articulosDTO.getImagen());
+        articulo.setNivelValido(articulosDTO.getNivelValido());
+        articulo.setDescripcion(articulosDTO.getDescripcion());
+        idEstado = estadoService.findById(articulosDTO.getIdEstado());
+        articulo.setEstado(idEstado);
+        //idCategoria = categoriaService.
+
+        articulosDAO.save(articulo);
         return "Se actualizo correctamente el articulo";
     }
 
     @Override
-    public String delete(Long idArticulo) throws Exception {
-        if(!articulosDAO.existsById(idArticulo)){
-            throw new Exception("El articulo no existe");
+    public String eliminar(Long idArticulo) throws Exception {
+        if(idArticulo == null){
+            throw new Exception("Se debe ingresar el id del articulo");
         }
+        if(!articulosDAO.existsById(idArticulo)){
+            throw new Exception("El articulo con id: " + idArticulo + " no existe");
+        }
+        //TODO: Falta validacion de si existe en articulo_adquirido
         articulosDAO.deleteById(idArticulo);
-        return "Se elimino correctamente el articulo";
+        return "El articulo se ha eliminado exitosamente";
+    }
+
+    @Override
+    public Articulos findById(Long idArticulo) throws Exception {
+        if(idArticulo == null){
+            throw new Exception("Se debe ingresar el id del articulo");
+        }
+        if(!articulosDAO.existsById(idArticulo)){
+            throw new Exception("El articulo con id: " + idArticulo + " no existe");
+        }
+        return articulosDAO.findById(idArticulo).get();
     }
 
     @Override
